Clean up with admin route
