+++ server/server.js
@@ -16,9 +16,7 @@ const ordersRouter = require('./routes/orders.router');
 
 const galleryRouter = require('./routes/gallery.router')
 
-// const adminRouter = require('./routes/admin.router')
 
-// const adminRouter = require('./routes/admin.router')
 
 // Body parser middleware
 app.use(bodyParser.json());
@@ -57,7 +55,7 @@ app.use('/orders', ordersRouter);
 
 app.use('/api/gallery', galleryRouter);
 
-// app.use('/api/admin', adminRouter)
+
 
 
 // Serve static files
